[CLEAN] - Moving Dependencies Around
This pull request is purely to move methods around to isolate the dependenicies a little better.

- moved anything related to kubernetes into pkg/utils/kubernetes package
- removed the unrequired factory method NewFactoryWithClient as it can be done with options
