+++ pkg/utils/kubernetes/time_test.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package utils
+package kubernetes
 
 import (
 	"testing"
